+++ WebJar.Application/ViewModel/AddOnViewModel.cs
@@ -1,4 +1,4 @@
 ﻿namespace WebJar.Application.ViewModel
 {
-    public record class AddOnViewModel(string Name, decimal Price);
+    public record class AddOnViewModel(string Name, long Price);
 }
