+++ WebJar.Application/ViewModel/DiscountViewModel.cs
@@ -1,4 +1,4 @@
 ﻿namespace WebJar.Application.ViewModel
 {
-    public record class DiscountViewModel(decimal? Amount, DateTime? ExpireDate);
+    public record class DiscountViewModel(long? Amount, DateTime? ExpireDate);
 }
