+++ WebJar.Application/ViewModel/PropertyValueViewModel.cs
@@ -1,5 +1,5 @@
 ﻿namespace WebJar.Application.ViewModel
 {
-    public record class PropertyValueViewModel(string Value, decimal Price,
+    public record class PropertyValueViewModel(string Value, long Price,
         string PriceType = "CONSTANT");
 }
